#191: PeerListScreen(component, modifier); presentation-layer rules
- PeerListScreen slims to (component, modifier): all callbacks built internally via
  rememberPeerCallbacks (private), which reads peerTransferComponent per row.
- PeerListContent drops hasPending/onPickerPick/autoSendEnabledFor params; tapAction
  uses callbacks.onClick (always peerComponent::onCardClick) or mobile-chooser sheet.
- RootContent shrinks to banner + Children switch; no inline callback wiring.
- presentation-layer.md: add rules — MutableStateFlow updates via update { }; components
  do not own domain logic (litmus test: would logic survive a UI rewrite?).
